requirements.txt
requests~=2.31.0
datetime~=5.2
python-dotenv~=1.0.0